Handle lack of responseText updates by Chrome
An onprogress event apparently does not guarantee that responseText
has been updated, and we may need to wait a little longer to retrieve
the data.  Hence, we reschedule a _parse_path() run for later.

This is known to happen on Chrome during a new path setup on an
otherwise idle stream.  Even though the onprogress event has fired,
responseText is empty so we cannot yet read the SYNC or NONCE frames.
It would appear that Chrome cannot be trusted to always fire
onprogress *after* data is available in responseText.