Protect against orphaned paths during recovery
An issue in Production highlighted the fact that individual tcstream
paths can become orphaned and cause all Mochiweb/Webmachine requests to
slow down as the number of orphaned paths grows.

The bug is that a path can wait forever in the receive in
tcst_path:wait_for_events/1, as we never do any server-side detection of
TCP connection shutdowns that originate from the browser--the current
code assumes that the server *always* initiates connection closure.

This fix ensures that the server will periodically try to send data over
the stream, which triggers Mochiweb/Webmachine's built-in TCP connection
shutdown detection.  This detection eliminates any orphaned paths.

This is a hasty fix that sends a character (`z') down an unused channel
name (`Z').  A proper fix would introduce a NOOP control message.